% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-functions.R
\docType{data}
\name{example_lasars_data}
\alias{example_lasars_data}
\title{Example dataset for lasars}
\format{
A data frame with 50 participants and 5 latent states:
\describe{
\item{subject}{Subject identifier}
\item{item}{Survey item identifier}
\item{response}{Chosen Likert scale response}
\item{subscale}{Latent state identifier}
\item{reverse}{Identifies positively- and reverse-scored items}
}
}
\source{
Open Source Psychometrics
}
\usage{
example_lasars_data
}
\description{
A small Big-5 dataset used in examples and vignettes.
}
\keyword{datasets}
