% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-functions.R
\name{gen_pp_data}
\alias{gen_pp_data}
\title{Generates posterior samples}
\usage{
gen_pp_data(
  sampled,
  n = 20,
  ll_func = sampled$ll_func,
  sampling_method = "pmwg",
  original_df = NULL,
  subject_colname = "subject",
  ...
)
}
\arguments{
\item{sampled}{A post-sampling object with parameter estimates. Should be the output of either run_lasars() or run_latstat()}

\item{n}{the number of posterior samples being generated}

\item{ll_func}{The likelihood function being used}

\item{sampling_method}{A character string which reflects the sampling approach to be taken. Can take the value 'mle' or 'pmwg'}

\item{original_df}{Necessary when sampling_method = 'mle'. Should be the original trial-wise data frame sampling was performed on}

\item{subject_colname}{A character string representing the subject identifier variable}

\item{...}{Additional parameters to pass into the pmwg run_stage calls}
}
\value{
A data frame
}
\description{
This function generates posterior-predictive data
}
