% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-functions.R
\name{lasars_ll_func}
\alias{lasars_ll_func}
\title{Likelihood function for lasars Model}
\usage{
lasars_ll_func(
  x,
  data,
  sample = FALSE,
  resp_opts,
  subject_colname,
  response_colname,
  subscale_colname,
  direction_colname = NULL,
  rev_score_id = NULL,
  sampling_method = "pmwg"
)
}
\arguments{
\item{x}{A named vector of parameter estimates}

\item{data}{A trial-wise data frame}

\item{sample}{A boolean variable reflecting whether the ll func should evaluate likelihood or generate data}

\item{resp_opts}{A numeric value representing the number of response options available in the scale}

\item{subject_colname}{A character string representing the subject identifier variable}

\item{response_colname}{A character string representing the chosen Likert scale response}

\item{subscale_colname}{A character string representing the latent state identifier variable}

\item{direction_colname}{A character string representing the reverse-coding variable}

\item{rev_score_id}{A character string representing the level of the direction variable which indicates reverse-scored items}

\item{sampling_method}{A character string which reflects the sampling approach to be taken.}
}
\value{
A likelihood value
}
\description{
This function is the powerhouse of the model
}
