% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-functions.R
\name{latstat_ll_func}
\alias{latstat_ll_func}
\title{Likelihood function for Latent State Only Model}
\usage{
latstat_ll_func(
  x,
  data,
  sample = FALSE,
  resp_opts,
  subject_colname,
  response_colname,
  latentState_colname,
  direction_colname,
  rev_score_id
)
}
\arguments{
\item{x}{A named vector of parameter estimates}

\item{data}{A trial-wise data frame}

\item{sample}{A boolean variable reflecting whether the ll func should evaluate likelihood or generate data}

\item{resp_opts}{A numeric value representing the number of response options available in the scale}

\item{subject_colname}{A character string representing the subject identifier variable}

\item{response_colname}{A character string representing the chosen Likert scale response}

\item{latentState_colname}{A character string representing the latent state identifier variable}

\item{direction_colname}{A character string representing the reverse-coding variable}

\item{rev_score_id}{A character string representing the level of the direction variable which indicates reverse-scored items}
}
\value{
A likelihood value
}
\description{
The powerhouse of the latstat model - calculates the likelihood of the observed data,
given the set of input parameters.
If sample = TRUE, input parameters will be used to generate data.
}
\examples{
\dontrun{
ll_func <- latstat_ll_func(x = c("mu.extra" = 0.2,
                                 "mu.open" = 0.1,
                                 "mu.consc" = 0.1,
                                 "mu.agree" = 0.2,
                                 "mu.neuro" = 0.1,
                                 "cut.1" = -0.3,
                                 "cut.2" = -0.2,
                                 "cut.4" = 0.5,
                                 sd = 0.3),
                           data = example_lasars_data,
                           resp_opts = 5,
                           subject_colname = "subject",
                           response_colname = "response",
                           latentState_colname = "subscale",
                           direction_colname = "reverse",
                           rev_score_id = "TRUE")
}
}
