% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-functions.R
\name{make_priors}
\alias{make_priors}
\title{Make pmwg priors}
\usage{
make_priors(
  data,
  resp_opts,
  latentState_colname,
  est_directPref = NULL,
  mu_prior = 0,
  centrePref_prior = 0,
  oddsPref_prior = 0,
  directionPref_prior = 0,
  threshold_prior = 0,
  diag_prior = 1,
  analysis = "lasars"
)
}
\arguments{
\item{data}{A trial-wise data frame}

\item{resp_opts}{A numeric value representing the number of response options available in the scale}

\item{latentState_colname}{A character string representing the latent state identifier variable}

\item{est_directPref}{A boolean value reflecting whether a directionPref parameter should be estimated}

\item{mu_prior}{A numeric value representing the prior for the mu parameter}

\item{centrePref_prior}{A numeric value representing the prior for the centrePref_prior parameter}

\item{oddsPref_prior}{A numeric value representing the prior for the oddsPref parameter}

\item{directionPref_prior}{A numeric value representing the prior for the directionPref parameter}

\item{threshold_prior}{A numeric value representing the prior for the thresholds in the latstat model}

\item{diag_prior}{A numeric value representing the prior for the diagonal}

\item{analysis}{A character string taking the value of 'lasars' or 'latstat' which indicates which model variant is being used}
}
\value{
A list of priors
}
\description{
Creates priors for pmwg sampling
}
\examples{
test_priors <- make_priors(data = example_lasars_data,
                           resp_opts = 5,
                           latentState_colname = "subscale",
                           est_directPref = TRUE,
                           mu_prior = 0.5,
                           centrePref_prior = 0.3)
test_priors

}
