% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-functions.R
\name{run_lasars}
\alias{run_lasars}
\title{Fit the lasars model}
\usage{
run_lasars(
  data,
  resp_opts,
  est_directPref,
  subject_colname,
  response_colname,
  latentState_colname,
  direction_colname = NULL,
  rev_score_id = NULL,
  priors = NULL,
  sampling_method = "pmwg",
  iterations = 3000,
  particles = 50,
  ...
)
}
\arguments{
\item{data}{A trial-wise data frame}

\item{resp_opts}{A numeric value representing the number of response options available in the scale}

\item{est_directPref}{A boolean value reflecting whether an directionPref parameter should be estimated}

\item{subject_colname}{A character string representing the subject identifier variable}

\item{response_colname}{A character string representing the chosen Likert scale response}

\item{latentState_colname}{A character string representing the latent state identifier variable}

\item{direction_colname}{A character string representing the reverse-coding variable}

\item{rev_score_id}{A character string representing the level of the direction variable which indicates reverse-scored items}

\item{priors}{A list of priors}

\item{sampling_method}{A character string which reflects the sampling approach to be taken.}

\item{iterations}{A numeric value reflecting the number of sampling iterations to run}

\item{particles}{A numeric value reflecting the number of proposed particles on each sampling iteration}

\item{...}{Additional parameters to pass into the pmwg run_stage calls}
}
\value{
A pmwg sampled object
}
\description{
Estimates latent-state and response-style parameters from
Likert-scale survey data.
}
\examples{
result <- run_lasars(
  data = example_lasars_data,
  resp_opts = 5,
  est_directPref = TRUE,
  subject_colname = "subject",
  response_colname = "response",
  latentState_colname = "subscale",
  direction_colname = "reverse",
  rev_score_id = "TRUE",
  iterations = 100,
  particles = 10
)


}
