% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-functions.R
\name{run_latstat}
\alias{run_latstat}
\title{Runs the latent state only model}
\usage{
run_latstat(
  data,
  resp_opts,
  subject_colname,
  response_colname,
  subscale_colname,
  direction_colname,
  rev_score_id,
  priors = NULL,
  sampling_method = "pmwg",
  iterations = 3000,
  particles = 50,
  ...
)
}
\arguments{
\item{data}{A trial-wise data frame}

\item{resp_opts}{A numeric value representing the number of response options available in the scale}

\item{subject_colname}{A character string representing the subject identifier variable}

\item{response_colname}{A character string representing the chosen Likert scale response}

\item{subscale_colname}{A character string representing the latent state identifier variable}

\item{direction_colname}{A character string representing the reverse-coding variable}

\item{rev_score_id}{A character string representing the level of the direction variable which indicates reverse-scored items}

\item{priors}{A list of priors}

\item{sampling_method}{A character string which reflects the sampling approach to be taken. Can take the value 'mle' or 'pmwg'.}

\item{iterations}{A numeric value reflecting the number of sampling iterations to run}

\item{particles}{A numeric value reflecting the number of proposed particles on each sampling iteration}

\item{...}{Additional parameters to pass into the pmwg run_stage calls}
}
\value{
A pmwg sampled object
}
\description{
This function runs the model over the data
}
